Clean up rustdoc passes
